Move WQ and TaskVine scaling into a
